+++ honeybadger/config.py
@@ -14,6 +14,7 @@ class Configuration(object):
         ('endpoint', str),
         ('params_filters', list),
         ('force_report_data', bool),
+        ('force_sync', bool)
     )
 
     def __init__(self, *args, **kwargs):
@@ -24,7 +25,8 @@ class Configuration(object):
         self.endpoint = 'https://api.honeybadger.io'
         self.params_filters = ['password', 'password_confirmation', 'credit_card', 'CSRF_COOKIE']
         self.force_report_data = False
-
+        self.force_sync = self.is_aws_lambda_environment
+        
         self.set_12factor_config()
         self.set_config_from_dict(kwargs)
 
@@ -59,3 +61,13 @@ class Configuration(object):
         :rtype: bool
         """
         return self.environment in self.DEVELOPMENT_ENVIRONMENTS
+
+    @property
+    def is_aws_lambda_environment(self):
+        """
+        Checks if you are in an AWS Lambda environment by checking for the existence
+        of "AWS_LAMBDA_FUNCTION_NAME" in the environment variables.
+
+        :rtype: bool
+        """
+        return os.environ.get("AWS_LAMBDA_FUNCTION_NAME") is not  None